Vulkan Rendering! (sorta) Fix vulkan validation layers not showing Add proper renderdoc support Fix bug where renderdoc would hang because of ScriptConsole update Begin supporting vulkan rendering * currently upsidedown * depth buffering is broken * colors seem to be smoothed Fix Vertex mapping to GLSL

+++ Engine/Public/Dusk/Vertex.hpp
@@ -19,7 +19,7 @@ struct Vertex
     glm::vec2 TexCoord1;
     glm::vec2 TexCoord2;
 
-    uint16_t Joints[4];
+    uint32_t Joints[4];
     float Weights[4];
     
 }; // struct Vertex
